Cast to avoid char-interpretation of uint8_t in logs.
The uint8_t in the log string is interpreted as a char, causing a
character to be logged if the loss is non-zero and terminates the string
with a '\0' in the zero case.

R=sprang@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/53449004

Cr-Commit-Position: refs/heads/master@{#9242}

+++ webrtc/video_engine/vie_encoder.cc
@@ -822,7 +822,7 @@ void ViEEncoder::OnNetworkChanged(uint32_t bitrate_bps,
                                   uint8_t fraction_lost,
                                   int64_t round_trip_time_ms) {
   LOG(LS_VERBOSE) << "OnNetworkChanged, bitrate" << bitrate_bps
-                  << " packet loss " << fraction_lost
+                  << " packet loss " << static_cast<int>(fraction_lost)
                   << " rtt " << round_trip_time_ms;
   DCHECK(send_payload_router_ != NULL);
   vcm_->SetChannelParameters(bitrate_bps, fraction_lost, round_trip_time_ms);
